License this package as Apache 2.0.
The license information in this package is still "TBD".
This commit is a proposal to use the Apache 2.0 license which is the
recommended license for ROS 2 packages.

Apache 2.0 is considered a permissive license with an express patent grant.
More information can be found at https://choosealicense.com/licenses/apache-2.0/ and https://apache.org/licenses/LICENSE-2.0