Remove local config directory and refactor helper.
Since we've removed the misra_rules.txt configuration file that's no
longer part of this package and setup.py does not need to reference it.

I've also proposed a small refactor to the replace_misra_rules helper to
allow the path to a local rules file to be specified rather than
assuming there is one distributed with the package.