Use a test fixture directory to store test sarif files.
These tests were using hard-coded paths instead of the path function at
the top of the file. That function was also reaching into the package
installation directory which may not be present or accessible during a
test run. Since the files are just for testing anyway, move them into a
test fixtures directory and use the current test file to locate the
fixtures directory relatively.

This also resolves the undeclared dependency on ament_index_python by
removing it.