Use a single set with tuple entries instead of nested dicts.
There isn't really a need to create a directory of nested dicts if all
we're concerned with is whether an exact result matching the ruleId,
artifact, and region has previously been seen so let's use a set with
tuple elements. It should be comparably performant to the nested hash
approach while being much more succinct.