This todo is resolved.
Since a colcon workspace has no *explicit* indicator there is no way I
know of to pre-emptively warn or or find the "workspace root" directory.
So it is now required that this command be run with the workspace root
as the current directory. The build and test command extraction
functions will fail if the current directory lacks a log/latest_build or
log/latest_test directory respectively which will enforce this
requirement.